Add usrmsg_users parameter
By default the emergency messages are sent to all logged in users
on the machine. rsyslog as a parameter to restrict these message
to a list of user. This new parameter permits to set this list
(to only root for instance). By default the parameter uses "*" (
all users) which was the previously hard coded value).

    https://github.com/scibian/puppet-module-saz-rsyslog/blob/scibian/stretch/debian/patches/0001-Add-usrmsg_users-parameter.patch